# fern/openapi/openapi.json
{
  "components": {
...
    "schemas": {
...
      "AgentSpec": {
        "description": "Agent Definition",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/RuntimeConfig"
          },
          "instructions": {
            "type": "string"
          },
          "mcp_servers": {
            "items": {
              "$ref": "#/components/schemas/MCPServer"
            },
            "type": "array"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/AgentSpecUserMessage"
            },
            "type": "array"
          },
          "model": {
            "$ref": "#/components/schemas/Model"
          },
          "response_format": {
            "$ref": "#/components/schemas/ResponseFormat"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillMount"
            },
            "type": "array"
          },
          "variables": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "required": [
          "model"
        ],
        "type": "object"
      },
      "AgentSpecUserMessage": {
        "properties": {
          "content": {
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "enum": [
              "user.message"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "content"
        ],
        "type": "object"
      },
...
      "AskUserQuestionsConfig": {
        "properties": {
          "enabled": {
            "default": true,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "ContextManagementConfig": {
        "default": {
          "compaction": {
            "enabled": true
          },
          "large_tool_response": {
            "enabled": true
          }
        },
        "properties": {
          "compaction": {
            "default": {
              "enabled": true
            },
            "properties": {
              "compaction_threshold_tokens": {
                "default": 50000,
                "exclusiveMinimum": 0,
                "type": "integer"
              },
              "enabled": {
                "default": true,
                "type": "boolean"
              }
            },
            "type": "object"
          },
          "large_tool_response": {
            "$ref": "#/components/schemas/LargeToolResponseConfig"
          }
        },
        "type": "object"
      },
      "CreateSessionRequest": {
        "properties": {
          "agent_spec": {
            "$ref": "#/components/schemas/AgentSpec"
          }
        },
        "required": [
          "agent_spec"
        ],
        "type": "object"
      },
...
      "DynamicSubAgentsConfig": {
        "properties": {
          "enabled": {
            "default": true,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "GenerativeUIConfig": {
        "properties": {
          "enabled": {
            "default": true,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "GetSessionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Session"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "LargeToolResponseConfig": {
        "default": {
          "enabled": true
        },
        "properties": {
          "enabled": {
            "default": true,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "ListSessionsOrder": {
        "default": "desc",
        "description": "Sort sessions by creation time. Defaults to \"desc\".",
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
      "ListSessionsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Session"
            },
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/TokenPagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "type": "object"
      },
...
      "MCPServer": {
        "properties": {
          "disable_tools": {
            "default": [],
            "items": {
              "anyOf": [
                {
                  "enum": [
                    "@all",
                    "@read-only"
                  ],
                  "type": "string"
                },
                {
                  "minLength": 1,
                  "type": "string"
                }
              ]
            },
            "type": "array"
          },
          "enable_tools": {
            "default": [
              "@all"
            ],
            "items": {
              "anyOf": [
                {
                  "enum": [
                    "@all",
                    "@read-only"
                  ],
                  "type": "string"
                },
                {
                  "minLength": 1,
                  "type": "string"
                }
              ]
            },
            "type": "array"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "preload": {
            "default": false,
            "type": "boolean"
          },
          "preload_tools": {
            "default": [],
            "items": {
              "anyOf": [
                {
                  "enum": [
                    "@all",
                    "@read-only"
                  ],
                  "type": "string"
                },
                {
                  "minLength": 1,
                  "type": "string"
                }
              ]
            },
            "type": "array"
          },
          "require_approval_for_tools": {
            "default": [
              "@write",
              "@destructive"
            ],
            "items": {
              "anyOf": [
                {
                  "enum": [
                    "@all",
                    "@write",
                    "@destructive"
                  ],
                  "type": "string"
                },
                {
                  "minLength": 1,
                  "type": "string"
                }
              ]
            },
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "Model": {
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "params": {
            "$ref": "#/components/schemas/ModelParams"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "ModelParams": {
        "properties": {
          "max_tokens": {
            "type": "number"
          },
          "parallel_tool_calls": {
            "type": "boolean"
          },
          "reasoning_effort": {
            "type": "string"
          },
          "temperature": {
            "type": "number"
          },
          "top_k": {
            "type": "number"
          },
          "top_p": {
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "RequestErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "message": {
                "type": "string"
              },
              "param": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "type": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ResponseFormat": {
        "oneOf": [
          {
            "properties": {
              "type": {
                "enum": [
                  "text"
                ],
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "properties": {
              "type": {
                "enum": [
                  "json_object"
                ],
                "type": "string"
              }
            },
            "required": [
              "type"
            ],
            "type": "object"
          },
          {
            "properties": {
              "json_schema": {
                "properties": {
                  "description": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "schema": {
                    "additionalProperties": {},
                    "type": "object"
                  },
                  "strict": {
                    "type": [
                      "boolean",
                      "null"
                    ]
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              },
              "type": {
                "enum": [
                  "json_schema"
                ],
                "type": "string"
              }
            },
            "required": [
              "type",
              "json_schema"
            ],
            "type": "object"
          }
        ]
      },
      "RuntimeConfig": {
        "properties": {
          "ask_user_questions": {
            "$ref": "#/components/schemas/AskUserQuestionsConfig"
          },
          "context_management": {
            "$ref": "#/components/schemas/ContextManagementConfig"
          },
          "dynamic_sub_agents": {
            "$ref": "#/components/schemas/DynamicSubAgentsConfig"
          },
          "generative_ui": {
            "$ref": "#/components/schemas/GenerativeUIConfig"
          },
          "iteration_limit": {
            "default": 100,
            "exclusiveMinimum": 0,
            "maximum": 1024,
            "type": "integer"
          },
          "sandbox": {
            "$ref": "#/components/schemas/SandboxConfig"
          }
        },
        "type": "object"
      },
      "SandboxConfig": {
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "file_downloads": {
            "default": true,
            "type": "boolean"
          },
          "network_policy": {
            "$ref": "#/components/schemas/SandboxNetworkPolicy"
          }
        },
        "required": [
          "enabled"
        ],
        "type": "object"
      },
...
      "SandboxNetworkPolicy": {
        "properties": {
          "auth_inject": {
            "items": {
              "properties": {
                "auth_data": {
                  "properties": {
                    "password": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "type": {
                      "enum": [
                        "basic"
                      ],
                      "type": "string"
                    },
                    "username": {
                      "pattern": "^\\S+$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "username",
                    "password"
                  ],
                  "type": "object"
                },
                "match": {
                  "properties": {
                    "hosts": {
                      "items": {
                        "minLength": 1,
                        "type": "string"
                      },
                      "minItems": 1,
                      "type": "array"
                    }
                  },
                  "required": [
                    "hosts"
                  ],
                  "type": "object"
                },
                "type": {
                  "enum": [
                    "git"
                  ],
                  "type": "string"
                }
              },
              "required": [
                "type",
                "match",
                "auth_data"
              ],
              "type": "object"
            },
            "maxItems": 1,
            "type": "array"
          }
        },
        "type": "object"
      },
      "Session": {
        "properties": {
          "agent_spec": {
            "$ref": "#/components/schemas/AgentSpec"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "agent_spec",
          "title",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "SkillMount": {
        "properties": {
          "description": {
            "description": "Concise guidance for when the agent should use the skill.",
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "description": "Name for the skill, used as directory name in sandbox.",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._-]+$",
            "type": "string"
          },
          "path": {
            "description": "Path to the skill directory within the repository. Omit to use the repository root.",
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._\\-/]+$",
            "type": "string"
          },
          "ref": {
            "description": "Git ref — branch name, tag, or commit SHA.",
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._\\-/]+$",
            "type": "string"
          },
          "type": {
            "enum": [
              "git"
            ],
            "type": "string"
          },
          "url": {
            "description": "Full HTTPS URL of a GitHub or GitLab repository.",
            "minLength": 1,
            "pattern": "^https:\\/\\/(github\\.com\\/[A-Za-z0-9._-]+\\/[A-Za-z0-9._-]+|gitlab\\.com\\/[A-Za-z0-9._-]+(?:\\/[A-Za-z0-9._-]+)+)(\\/|\\.git)?$",
            "type": "string"
          }
        },
        "required": [
          "type",
          "url",
          "name",
          "description",
          "ref"
        ],
        "type": "object"
      },
...
      "TokenPagination": {
        "properties": {
          "limit": {
            "exclusiveMinimum": 0,
            "type": "integer"
          },
          "next_page_token": {
            "type": "string"
          },
          "previous_page_token": {
            "type": "string"
          }
        },
        "required": [
          "limit"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Agent server exposing models, MCP servers and skills from local YAML config.",
    "title": "Agent Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "description": "List sessions (newest first by default), token-paginated. Pass `page_token` to fetch the next page, keeping the other query params constant.",
        "parameters": [
          {
            "description": "Page size. Defaults to 10, max 100.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "description": "Page size. Defaults to 10, max 100.",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Sort sessions by creation time. Defaults to \"desc\".",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ListSessionsOrder"
            }
          },
          {
            "description": "Opaque token from a previous response `next_page_token`.",
            "in": "query",
            "name": "page_token",
            "required": false,
            "schema": {
              "description": "Opaque token from a previous response `next_page_token`.",
              "type": "string"
            }
          },
          {
            "description": "Inclusive lower bound on `created_at` (ISO-8601 / RFC 3339).",
            "in": "query",
            "name": "start_timestamp",
            "required": false,
            "schema": {
              "description": "Inclusive lower bound on `created_at` (ISO-8601 / RFC 3339).",
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Inclusive upper bound on `created_at` (ISO-8601 / RFC 3339).",
            "in": "query",
            "name": "end_timestamp",
            "required": false,
            "schema": {
              "description": "Inclusive upper bound on `created_at` (ISO-8601 / RFC 3339).",
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSessionsResponse"
                }
              }
            },
            "description": "Paginated sessions."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "Invalid query parameters or page token."
          }
        },
        "summary": "List sessions",
        "tags": [
          "Sessions"
        ],
        "x-fern-pagination": {
          "cursor": "$request.page_token",
          "next_cursor": "$response.pagination.next_page_token",
          "results": "$response.data"
        },
        "x-fern-sdk-group-name": [
          "sessions"
        ],
        "x-fern-sdk-method-name": "list"
      },
      "post": {
        "description": "Create a session holding an inline agent spec. Turns are executed against this spec.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSessionResponse"
                }
              }
            },
            "description": "Session created."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "Invalid request body."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorResponse"
                }
              }
            },
            "description": "The agent spec is valid but requires a capability this server does not provide (e.g. sandbox or skills)."
          }
        },
        "summary": "Create a session",
        "tags": [
          "Sessions"
        ],
        "x-fern-sdk-group-name": [
          "sessions"
        ],
        "x-fern-sdk-method-name": "create"
      }
    },
...
  },
...
}